+++ packages/f-checkout/test-utils/component-objects/f-checkout.component.js
@@ -1,26 +1,68 @@
-const checkoutComponent = () => $('[data-test-id="checkout-component"]');
+/* global browser, $ */
+
+import {
+    ALLERGEN_LINK,
+    CHECKOUT_COMPONENT,
+    ORDER_TIME_DROPDOWN,
+    ORDER_TIME_DROPDOWN_OPTIONS,
+    USER_NOTE_INPUT,
+    GO_TO_PAYMENT_BUTTON,
+    FIELDS
+} from './f-checkout-selectors';
+
+const { doesElementExist } = require('../../../../test/utils/webdriverio-extensions')(browser);
+
+const checkoutComponent = () => $(CHECKOUT_COMPONENT);
+
+// Dropdown
+
+const orderTimeDropdown = () => $(ORDER_TIME_DROPDOWN);
+const orderTimeDropdownOptions = () => $$(ORDER_TIME_DROPDOWN_OPTIONS);
+
+// Buttons
+
+const allergenLink = () => $(ALLERGEN_LINK);
+const goToPaymentButton = () => $(GO_TO_PAYMENT_BUTTON);
 
 // Form Fields
 
-const userNoteInput = () => $('[data-test-id="user-note"] textarea');
+const userNoteInput = () => $(USER_NOTE_INPUT);
 
-const allergenLink = () => $('[data-test-id="allergy-button"]');
-
-const fulfillmentTimeDropdown = () => $('[data-test-id="fulfillment-time"]');
-
-const goToPaymentButton = () => $('[data-test-id="confirm-payment-submit-button"]');
-
-exports.inputs = {
-    mobileNumber: () => $('[data-test-id="input-mobile-number"]'),
-    addressLine1: () => $('[data-test-id="input-address-line-1"]'),
-    addressLine2: () => $('[data-test-id="input-address-line-2"]'),
-    addressCity: () => $('[data-test-id="input-address-city"]'),
-    addressPostcode: () => $('[data-test-id="input-address-postcode"]')
+const fields = {
+    mobileNumber: {
+        input: () => $(FIELDS.mobileNumber.input),
+        error: () => $(FIELDS.mobileNumber.error)
+    },
+    addressLine1: {
+        input: () => $(FIELDS.addressLine1.input),
+        error: () => $(FIELDS.addressLine1.error)
+    },
+    addressLine2: {
+        input: () => $(FIELDS.addressLine2.input),
+        error: ''
+    },
+    addressCity: {
+        input: () => $(FIELDS.addressCity.input),
+        error: () => $(FIELDS.addressCity.error)
+    },
+    addressPostcode: {
+        input: () => $(FIELDS.addressPostcode.input),
+        error: () => $(FIELDS.addressPostcode.error)
+    }, 
+    userNote: {
+        input: () => $(FIELDS.userNote.input), 
+        error: ''
+    }
 };
 
+exports.isFieldErrorDisplayed = fieldName => fields[fieldName].error().isDisplayed();
+exports.isFieldDisplayed = fieldName => fields[fieldName].input().isDisplayed();
 exports.waitForCheckoutComponent = () => checkoutComponent().waitForExist();
 exports.isCheckoutComponentDisplayed = () => checkoutComponent().isDisplayed();
 exports.isAllergenLinkDisplayed = () => allergenLink().isDisplayed();
+exports.isOrderTimeDropdownDisplayed = () => orderTimeDropdown().isDisplayed();
+exports.userNoteMaxCharacterCount = () => userNoteInput().getAttribute('maxlength');
+exports.clickPaymentButton = () => goToPaymentButton().click();
 
 /**
  * @description
@@ -32,40 +74,61 @@ exports.isAllergenLinkDisplayed = () => allergenLink().isDisplayed();
  * @param {String} addressInfo.line2 Second line of the user's address
  * @param {String} addressInfo.city City of the user's address
  * @param {String} addressInfo.postcode Postcode of the user's address
+ * @param {String} addressInfo.note The user's extra note
  */
-exports.submitCheckoutForm = addressInfo => {
+exports.populateCheckoutForm = addressInfo => {
     exports.waitForCheckoutComponent();
-    mobileNumberInput().setValue(addressInfo.mobileNumber);
-    addressLine1Input().setValue(addressInfo.line1);
-    addressLine2Input().setValue(addressInfo.line2);
-    addressCityInput().setValue(addressInfo.city);
-    addressPostcodeInput().setValue(addressInfo.postcode);
+    fields.mobileNumber.input().setValue(addressInfo.mobileNumber);
+    fields.addressLine1.input().setValue(addressInfo.line1);
+    fields.addressLine2.input().setValue(addressInfo.line2);
+    fields.addressCity.input().setValue(addressInfo.city);
+    fields.addressPostcode.input().setValue(addressInfo.postcode);
+    fields.userNote.input().setValue(addressInfo.note);
 };
-
 /**
  * @description
- * Sets the value of the fulfillment time dropdown based on visible text.
- *
- * @param {String} fulfillmentTimeText The text visible text value of the fulfillment time
+ * Sets the value of the order time in dropdown based on visible text.
+ * 
+ * @param {String} orderTime The visible text value of the order time
  */
-exports.selectFulfillmentTime = fulfillmentTimeText => {
-    fulfillmentTimeDropdown().selectByVisibleText(fulfillmentTimeText);
+exports.selectOrderTime = orderTime => {
+    orderTimeDropdown().selectByVisibleText(orderTime);
 };
-
+/** 
+ * @description
+ * The time of the order should increase when a higher index is applied.
+ * 
+ * @param {Number} index The index of the `orderTimeDropdownOptions` array
+ */
+exports.getOrderTimeOptionText = (index) => {
+    return orderTimeDropdownOptions()[index].getText();
+};
 /**
  * @description
  * Sets the value of the user note.
- *
- * @param {String} userNote The user note value to be entered
+ * 
+ * @param {Object} addressInfo
+ * @param {String} addressInfo.note The user's extra note
  */
-exports.inputUserNote = userNote => {
-    userNoteInput().setValue(userNote);
+exports.inputUserNote = addressInfo => {
+    fields.userNote.input().setValue(addressInfo.note);
 };
-
 /**
  * @description
- * Submits the checkout form.
+ * Grabs the length of characters of the user note.
+ * 
+ * @returns {number} The length of the user note
+ */
+exports.getUserNoteLength = () => {
+    return userNoteInput().getValue().length;
+};
+/**
+ * @description
+ *Submit the checkout form.
  */
 exports.submit = () => {
     goToPaymentButton().click();
 };
+
+exports.doesErrorMessageExist = errorMessage => doesElementExist(FIELDS[errorMessage].error);
+exports.doesInputFieldExist = inputField => doesElementExist(FIELDS[inputField].input);