+++ packages/f-form-field/src/components/FormField.vue
@@ -3,10 +3,11 @@
         :data-theme-formfield="theme"
         :class="[
             $style['c-formField'], {
-                [$style['c-formField--invalid']]: hasError
+                [$style['c-formField--invalid']]: hasError,
+                [$style['c-formField--grouped']]: isGrouped
             }
         ]"
-        data-test-id="form-field-component">
+        :data-test-id="testId.container">
         <div
             :class="$style['c-formField-inputWrapper']">
             <form-label
@@ -14,20 +15,19 @@
                 :label-style="normalisedLabelStyle"
                 :for="uniqueId"
                 :is-inline="isInline"
-                data-test-id="form-field-label">
+                :data-test-id="testId.label">
                 {{ labelText }}
             </form-label>
 
             <form-dropdown
                 v-if="isDropdown"
                 :id="`${uniqueId}`"
-                v-bind="$attrs"
+                :attributes="$attrs"
                 :type="normalisedInputType"
-                :data-test-id="testId"
                 :class="[
-                    $style['o-form-field'],
                     $style['c-formField-input'],
-                    $style['c-formField-input-inputFields--focus']
+                    $style['c-formField-dropdownContainer'],
+                    $style['c-formField-input--focus']
                 ]"
                 :dropdown-options="dropdownOptions"
                 @update="updateOption"
@@ -40,13 +40,10 @@
                 v-bind="$attrs"
                 :type="normalisedInputType"
                 placeholder=" "
-                :data-test-id="testId"
+                :data-test-id="testId.input"
                 :class="[
-                    $style['o-form-field'],
                     $style['c-formField-input'],
-                    $style['c-formField-input-textField'], {
-                        [$style['c-formField-input-inputFields--focus']]: isInputField
-                    }
+                    (isSelectionControl ? $style['c-formField-input--focus'] : '')
                 ]"
                 @input="updateValue"
                 v-on="listeners"
@@ -56,7 +53,7 @@
                 :label-style="normalisedLabelStyle"
                 :for="uniqueId"
                 :is-inline="isInline"
-                data-test-id="form-field-label--inline">
+                :data-test-id="`${testId.label}--inline`">
                 {{ labelText }}
             </form-label>
         </div>
@@ -121,14 +118,14 @@ export default {
             default: false
         },
 
-        dataTestId: {
-            type: String,
-            default: ''
-        },
-
         dropdownOptions: {
             type: Array,
             default: () => null
+        },
+
+        isGrouped: {
+            type: Boolean,
+            default: false
         }
     },
 
@@ -179,7 +176,13 @@ export default {
         },
 
         testId () {
-            return this.dataTestId || this.$attrs.name || false;
+            const formFieldName = this.$attrs.name;
+
+            return {
+                container: formFieldName ? `formfield-${formFieldName}` : 'formfield-container',
+                input: formFieldName ? `formfield-${formFieldName}-input` : 'formfield-input',
+                label: formFieldName ? `formfield-${formFieldName}-label` : 'formfield-label'
+            };
         },
 
         isInline () {
@@ -191,7 +194,7 @@ export default {
             return this.inputType === 'dropdown';
         },
 
-        isInputField () {
+        isSelectionControl () {
             return !(this.inputType === 'radio' || this.inputType === 'checkbox');
         }
     },
@@ -245,34 +248,24 @@ $form-input-focus                         : $blue--light;
         margin-top: spacing(x2);
     }
 }
+
     .c-formField-inputWrapper {
         position: relative;
     }
 
-    .c-formField-input-textField {
-        padding: $form-input-padding;
-    }
-
-    .c-formField-input-inputFields--focus {
-        &:focus,
-        &:active,
-        &:focus-within {
-            box-shadow: 0 0 0 2pt $form-input-focus;
-            outline: none;
-        }
-    }
-
     .c-formField-input {
         width: 100%;
-        @include rem(height, $form-input-height); //convert height to rem
-        @include font-size($form-input-fontSize);
         font-family: $font-family-base;
-        color: $form-input-textColour;
+        @include font-size($form-input-fontSize);
         font-weight: $font-weight-base;
+        color: $form-input-textColour;
+        @include rem(height, $form-input-height); //convert height to rem
+
         background-color: $form-input-bg;
         border: $form-input-borderWidth solid $form-input-borderColour;
         border-radius: $form-input-borderRadius;
         background-clip: padding-box;
+        padding: $form-input-padding;
 
         &:hover {
             background-color: $form-input-bg--hover;
@@ -295,4 +288,39 @@ $form-input-focus                         : $blue--light;
         }
     }
 
+    .c-formField-input--focus {
+        &:focus,
+        &:active,
+        &:focus-within {
+            box-shadow: 0 0 0 2pt $form-input-focus;
+            outline: none;
+        }
+    }
+
+    .c-formField-dropdownContainer {
+        padding: 0;
+    }
+
+    .c-formField--grouped {
+        & + & {
+            margin-top: 0;
+
+            .c-formField-input {
+                border-radius: 0 0 $form-input-borderRadius $form-input-borderRadius;
+                border-top: 0;
+            }
+        }
+    }
+
+    .c-formField--grouped:not(:last-child) {
+        .c-formField-input {
+            border-radius: 0;
+        }
+    }
+
+    .c-formField--grouped:first-child {
+        .c-formField-input {
+            border-radius: $form-input-borderRadius $form-input-borderRadius 0 0;
+        }
+    }
 </style>