+++ packages/f-form-field/stories/FormField.stories.js
@@ -17,7 +17,7 @@ export const FormFieldComponent = () => ({
             default: select('Locale', ['en-GB', 'en-AU'])
         },
         labelText: {
-            default: text('Label Text', 'First Name')
+            default: text('Label Text', 'First name')
         },
         inputType: {
             default: select('Input Type', VALID_INPUT_TYPES.concat(CUSTOM_INPUT_TYPES))
@@ -30,13 +30,16 @@ export const FormFieldComponent = () => ({
         },
         dropdownOptions: {
             default: array('Dropdown Options', ['As soon as possible', 'Today in 5 minutes'], ',')
+        },
+        isGrouped: {
+            default: boolean('isGrouped', false)
         }
     },
     parameters: {
         notes: 'some documentation here'
     },
     template:
-        '<form-field :locale="locale" :labelText="labelText" :inputType="inputType" :labelStyle="labelStyle" :hasError="hasError" :dropdownOptions="dropdownOptions"/>'
+        '<form-field :locale="locale" :labelText="labelText" :inputType="inputType" :labelStyle="labelStyle" :hasError="hasError" :dropdownOptions="dropdownOptions" :isGrouped="isGrouped" />'
 });
 
 FormFieldComponent.storyName = 'f-form-field';